Payment Backend

+++ PlatformApi/Dtos/Response/PlanPaiementResponseDto.cs
@@ -11,7 +11,7 @@ namespace PlatformApi.Dtos.Response
 
         public string description { get; set; }
 
-        public float prix { get; set; }
+        public decimal prix { get; set; }
 
         [JsonIgnore]
         public IList<Paiement> Paiements { get; set; }
